+++ core/src/main/java/react4j/internal/arez/package-info.java
@@ -1,5 +1,0 @@
-/**
- * React4j internal support code for Arez aspect.
- * Code under this package can change at any time and should not be referenced outside this package.
- */
-package react4j.internal.arez;