+++ core/src/main/java/react4j/internal/package-info.java
@@ -1,5 +1,5 @@
 /**
  * React4j internal support code.
- * Code under this package can change at any time and should not be referenced outside this package.
+ * Code under this package can change at any time and should not be referenced outside this package except in generated code.
  */
 package react4j.internal;