+++ core/src/test/java/react4j/test/JDependTest.java
@@ -27,7 +27,6 @@ public class JDependTest
 
     final JavaPackage react4j = constraint.addPackage( "react4j" );
     final JavaPackage react4jInternal = constraint.addPackage( "react4j.internal" );
-    final JavaPackage react4jInternalArez = constraint.addPackage( "react4j.internal.arez" );
     constraint.addPackage( "react4j.annotations" );
     final JavaPackage braincheck = constraint.addPackage( "org.realityforge.braincheck" );
     final JavaPackage jsinteropAnnotations = constraint.addPackage( "jsinterop.annotations" );
@@ -45,14 +44,11 @@ public class JDependTest
 
     react4jInternal.dependsUpon( jsinteropAnnotations );
     react4jInternal.dependsUpon( jsinteropBase );
+    react4jInternal.dependsUpon( elemental2Promise );
     react4jInternal.dependsUpon( elemental2Core );
     react4jInternal.dependsUpon( react4j );
-
-    react4jInternalArez.dependsUpon( jsinteropBase );
-    react4jInternalArez.dependsUpon( elemental2Promise );
-    react4jInternalArez.dependsUpon( arez );
-    react4jInternalArez.dependsUpon( arezSpy );
-    react4jInternalArez.dependsUpon( elemental2Core );
+    react4jInternal.dependsUpon( arez );
+    react4jInternal.dependsUpon( arezSpy );
 
     final DependencyConstraint.MatchResult result = jdepend.analyzeDependencies( constraint );