+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Introduce a `@Render` annotation that is used to annotate the render method of a component. This makes it possible to completely remove the `react.Component` class.
 * Rename the `@ReactComponent` annotation to `@View` and update all supporting infrastructure to use the new terminology.
 * Remove the GWT module named `react4j.internal.arez.Arez` as it was never used and all the source in the package included when the `react4j.React` module was inherited.
+* Merge the `react4j.internal.arez` package into `react4j.internal` as the contents of these packages are used by all component types.
 
 ### [v0.161](https://github.com/react4j/react4j/tree/v0.161) (2020-06-02) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.160...v0.161) · [API Differences](https://react4j.github.io/api-diff?key=react4j&old=0.160&new=0.161)